Modified title, legend, and float data divs:
The various labeling div's (title, legend, float data box) have
gotten a little more elaborate, so I moved the initial code for
setting those up to plain html outside the main script, and labeled
each with id's for easier targeting by d3.

Also added a new onclick behavior to the float circles/paths so that
hover only displays the float name in the data box, while a click
supplies the full detail.  This avoids the problem of accidental
hovering in crowded areas when trying to navigate the mouse over
to the download links (or anywhere else).